+++ examples/ping-pong/src/main.rs
@@ -2,7 +2,6 @@ mod basic;
 mod retry;
 
 use std::collections::HashSet;
-use std::env;
 use std::io::Write;
 
 use assertables::assume;
@@ -48,9 +47,6 @@ struct Args {
 
 fn main() {
     let args = Args::parse();
-    if args.impl_path.ends_with(".py") {
-        env::set_var("PYTHONPATH", "../../python");
-    }
     init_logger(LevelFilter::Debug);
     let config = TestConfig {
         impl_path: args.impl_path,